change package to allow bw compatibility
