+++ modules/client-java/src/main/java/be/wegenenverkeer/atomium/client/FeedFetchException.java
@@ -1,0 +1,9 @@
+package be.wegenenverkeer.atomium.client;
+
+public class FeedFetchException extends RuntimeException {
+
+    public FeedFetchException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+}
