+++ modules/client-java/src/main/java/be/wegenenverkeer/atomium/client/FeedPositionStrategy.java
@@ -1,0 +1,7 @@
+package be.wegenenverkeer.atomium.client;
+
+import io.reactivex.rxjava3.core.Single;
+
+public interface FeedPositionStrategy {
+    <E> Single<FeedPosition> getNextFeedPosition(CachedFeedPage<E> feedPage);
+}
