+++ modules/client-java/src/main/java/be/wegenenverkeer/atomium/client/FeedPositionStrategyFrom.java
@@ -1,0 +1,14 @@
+package be.wegenenverkeer.atomium.client;
+
+public class FeedPositionStrategyFrom extends DefaultFeedPositionStrategy {
+    private final FeedPosition feedPosition;
+
+    public FeedPositionStrategyFrom(FeedPosition feedPosition) {
+        this.feedPosition = feedPosition;
+    }
+
+    @Override
+    protected <E> FeedPosition initialFeedPosition(CachedFeedPage<E> page) {
+        return feedPosition;
+    }
+}
