+++ modules/client-java/src/main/java/be/wegenenverkeer/atomium/client/FeedPositionStrategyFromNowOn.java
@@ -1,0 +1,8 @@
+package be.wegenenverkeer.atomium.client;
+
+public class FeedPositionStrategyFromNowOn extends DefaultFeedPositionStrategy {
+    @Override
+    protected <E> FeedPosition initialFeedPosition(CachedFeedPage<E> page) {
+        return new FeedPosition(page.getSelfHref(), page.getMostRecentEntryId());
+    }
+}
