+++ modules/client-java/src/main/java/be/wegenenverkeer/atomium/client/FeedProcessingException.java
@@ -1,0 +1,12 @@
+package be.wegenenverkeer.atomium.client;
+
+public class FeedProcessingException extends RuntimeException {
+
+    final String entryId;
+
+    public FeedProcessingException(String entryId, String message) {
+        super(message);
+        this.entryId = entryId;
+    }
+
+}
