+++ modules/client-java/src/main/java/be/wegenenverkeer/atomium/client/PageFetcher.java
@@ -1,0 +1,12 @@
+package be.wegenenverkeer.atomium.client;
+
+import io.reactivex.rxjava3.core.Single;
+
+import java.util.Optional;
+
+public interface PageFetcher<E> {
+    Single<CachedFeedPage<E>> fetch(String url, Optional<String> etag);
+    void close();
+    Class<E> getEntryTypeMarker();
+    RetryStrategy getRetryStrategy();
+}
