+++ modules/client-java/src/main/java/be/wegenenverkeer/atomium/client/rxhttpclient/ClientRequestCustomizer.java
@@ -1,0 +1,7 @@
+package be.wegenenverkeer.atomium.client.rxhttpclient;
+
+import be.wegenenverkeer.rxhttpclient.ClientRequestBuilder;
+
+public interface ClientRequestCustomizer {
+    void apply(ClientRequestBuilder builder);
+}
