--- modules/client-java/src/main/java/be/wegenenverkeer/atomium/japi/client/FeedFetchException.java
@@ -1,9 +1,0 @@
-package be.wegenenverkeer.atomium.japi.client;
-
-public class FeedFetchException extends RuntimeException {
-
-    public FeedFetchException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-}
