--- modules/client-java/src/main/java/be/wegenenverkeer/atomium/japi/client/FeedPosition.java
@@ -1,24 +1,0 @@
-package be.wegenenverkeer.atomium.japi.client;
-
-public class FeedPosition {
-    String pageUrl;
-    String entryId;
-
-    public FeedPosition(String pageUrl, String entryId) {
-        this.pageUrl = pageUrl;
-        this.entryId = entryId;
-    }
-
-    public FeedPosition(String pageUrl) {
-        this.pageUrl = pageUrl;
-    }
-
-    public String getPageUrl() {
-        return pageUrl;
-    }
-
-    public String getEntryId() {
-        return entryId;
-    }
-}
-
