--- modules/client-java/src/main/java/be/wegenenverkeer/atomium/japi/client/FeedPositionStrategy.java
@@ -1,7 +1,0 @@
-package be.wegenenverkeer.atomium.japi.client;
-
-import io.reactivex.rxjava3.core.Single;
-
-public interface FeedPositionStrategy {
-    <E> Single<FeedPosition> getNextFeedPosition(CachedFeedPage<E> feedPage);
-}
