--- modules/client-java/src/main/java/be/wegenenverkeer/atomium/japi/client/FeedPositionStrategyFrom.java
@@ -1,14 +1,0 @@
-package be.wegenenverkeer.atomium.japi.client;
-
-public class FeedPositionStrategyFrom extends DefaultFeedPositionStrategy {
-    private final FeedPosition feedPosition;
-
-    public FeedPositionStrategyFrom(FeedPosition feedPosition) {
-        this.feedPosition = feedPosition;
-    }
-
-    @Override
-    protected <E> FeedPosition initialFeedPosition(CachedFeedPage<E> page) {
-        return feedPosition;
-    }
-}
