--- modules/client-java/src/main/java/be/wegenenverkeer/atomium/japi/client/FeedPositionStrategyFromStart.java
@@ -1,8 +1,0 @@
-package be.wegenenverkeer.atomium.japi.client;
-
-public class FeedPositionStrategyFromStart extends DefaultFeedPositionStrategy {
-    @Override
-    protected <E> FeedPosition initialFeedPosition(CachedFeedPage<E> page) {
-        return new FeedPosition(page.getLastHref());
-    }
-}
