--- modules/client-java/src/main/java/be/wegenenverkeer/atomium/japi/client/FeedProcessingException.java
@@ -1,12 +1,0 @@
-package be.wegenenverkeer.atomium.japi.client;
-
-public class FeedProcessingException extends RuntimeException {
-
-    final String entryId;
-
-    public FeedProcessingException(String entryId, String message) {
-        super(message);
-        this.entryId = entryId;
-    }
-
-}
