--- modules/client-java/src/main/java/be/wegenenverkeer/atomium/japi/client/rxhttpclient/ClientRequestCustomizer.java
@@ -1,7 +1,0 @@
-package be.wegenenverkeer.atomium.japi.client.rxhttpclient;
-
-import be.wegenenverkeer.rxhttpclient.ClientRequestBuilder;
-
-public interface ClientRequestCustomizer {
-    void apply(ClientRequestBuilder builder);
-}
